# api-docs.yml
openapi: 3.1.0
info:
  description: |
    2024 멋쟁이사자처럼 중앙해커톤 남송리3번지팀 데모 API 명세
  version: 0.1.0
  title: stepper
  contact:
    name: zionhann
    email: hanzion1108@gmail.com
servers:
  - url: https://api.zionhann.com/stepper
    description: production
  - url: http://localhost:8080
    description: local
security:
  - X-CSRF-TOKEN: [ ]
...
paths:
...
  /v1/goals/{goalId}/chats:
    post:
      tags:
        - chats
      summary: 특정 목표에 대해 AI로 일지 작성 채팅방 생성
      description: " "
      operationId: initChat
      parameters:
        - name: goalId
          in: path
          description: 목표 ID
          required: true
          schema:
            type: number
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
              examples:
                ChatResponseExample:
                  value:
                    chatId: thread_StEPpEr
  /v1/chats/{chatId}/history:
    get:
      tags:
        - chats
      summary: 특정 채팅방의 대화내역 조회
      description: " "
      operationId: getChatMessages
      parameters:
        - name: chatId
          in: path
          description: 채팅방 ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChatHistoryResponse'
              examples:
                ChatHistoryResponseExample:
                  value:
                    messages:
                      - role: CHATBOT
                        content: "안녕하세요 😊\n저는 당신의 성장 여정을 함께할 AI (챗봇 이름)입니다 :)"
                      - role: CHATBOT
                        content: "자신의 생각과 감정을 솔직하게 기록하면서, 매일 조금씩 더 나은 나를 발견해 보세요.\n오늘의 기록이 내일의 당신을 더욱 빛나게 할 거예요.\n자, 이제 시작해 볼까요? 당신의 이야기를 들려주세요!"
                      - role: CHATBOT
                        content: "어떤 계기로 ‘중앙해커톤 우승'이라는 목표를 가지게 되었나요?"
                      - role: USER
                        content: "한동 멋사에서 아기사자로 활동하는 동안 우승하고 싶어!"
  /v1/chats/{chatId}/summary:
    post:
      tags:
        - chats
      summary: 특정 채팅방의 대화 내용 요약 (일지 초안 작성)
      description: " "
      operationId: chatSummary
      parameters:
        - name: chatId
          in: path
          description: 채팅방 ID
          required: true
          schema:
            type: number
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatSummaryResponse'
              examples:
                ChatSummaryResponseExample:
                  value:
                    content: "중앙해커톤에 참여하면서 '우승'이라는 목표를 세웠다. 활동 기간 동안 우승하고 싶다는 열망이 커졌고, 이를 위해 매주 두 번씩 모여 서비스에 대한 회의를 진행했다. 지속적으로 피드백을 주고받고, AI 챗봇 기능이 정말 중요하다는 이야기를 듣고 반드시 선보여야겠다는 다짐을 했다.\n\nAI 챗봇 기능의 발전을 위해 기능 플로우에 대한 시퀀스 다이어그램을 만들고 순차적으로 구현해 나갔다."
components:
  securitySchemes:
    X-CSRF-TOKEN:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
  schemas:
...
    ChatResponse:
      type: object
      properties:
        chatId:
          type: string
          description: 채팅방 ID
    ChatHistoryResponse:
      type: object
      properties:
        role:
          type: string
          description: 메시지 주체
        content:
          type: string
          description: 메시지 내용
    ChatSummaryResponse:
      type: object
      properties:
        content:
          type: string
          description: 요약한 내용